stage 2:
- move functionality from server.py to seperate modules divded by their area
- move tests from tests_server.py where applicable right now to get it to run
