+++ grizzly-ls/src/grizzly_ls/constants.py
@@ -1,0 +1,4 @@
+FEATURE_INSTALL: str = 'grizzly-ls/install'
+
+MARKER_LANGUAGE: str = '# language:'
+MARKER_NO_STEP_IMPL: str = 'No step implementation found for:'
